refactor: NotificationHandler 패키지 이동
